RD-4423: Adjusted translation labels and refactored types
