+++ widgets/tokens/src/consts.ts
@@ -1,0 +1,1 @@
+export const widgetTranslationPath = 'widgets.tokens';
